+++ lib/recording_studio_user/engine.rb
@@ -20,10 +20,12 @@ module RecordingStudioUser
       ActiveSupport.on_load(:action_controller) do
         helper RecordingStudioUser::OmniauthHelper if respond_to?(:helper)
         helper RecordingStudioUser::AuthRoutesHelper if respond_to?(:helper)
+        helper RecordingStudioUser::AuthSiteLogoHelper if respond_to?(:helper)
       end
       ActiveSupport.on_load(:action_view) do
         include RecordingStudioUser::OmniauthHelper
         include RecordingStudioUser::AuthRoutesHelper
+        include RecordingStudioUser::AuthSiteLogoHelper
       end
     end