+++ test/dummy/app/models/admin_root.rb
@@ -7,5 +7,6 @@ class AdminRoot < ApplicationRecord
   recording_studio_admin_sections do
     section :root
     section :users
+    section :site_settings
   end
 end