+++ test/dummy/db/migrate/20260904060001_add_depends_on_recording_id_to_recording_studio_accesses.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class AddDependsOnRecordingIdToRecordingStudioAccesses < ActiveRecord::Migration[8.1]
+  def change
+    return if column_exists?(:recording_studio_accesses, :depends_on_recording_id)
+
+    add_column :recording_studio_accesses, :depends_on_recording_id, :uuid
+    add_index :recording_studio_accesses, :depends_on_recording_id
+  end
+end